@@ -11,32 +11,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.16",
-    "@angular/cdk": "~10.0.2",
-    "@angular/common": "~12.2.16",
-    "@angular/compiler": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.16",
-    "@angular/material": "^10.0.2",
-    "@angular/material-moment-adapter": "^10.0.2",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/router": "~12.2.16",
+    "@angular/animations": "~14.0.6",
+    "@angular/cdk": "~11.2.13",
+    "@angular/common": "~14.0.6",
+    "@angular/compiler": "~14.0.6",
+    "@angular/core": "~14.0.6",
+    "@angular/forms": "~14.0.6",
+    "@angular/material": "^11.2.13",
+    "@angular/material-moment-adapter": "^11.2.13",
+    "@angular/platform-browser": "~14.0.6",
+    "@angular/platform-browser-dynamic": "~14.0.6",
+    "@angular/router": "~14.0.6",
     "@ngrx/effects": "^12.5.1",
     "@ngrx/entity": "^12.5.1",
     "@ngrx/store": "^12.5.1",
     "@ngrx/store-devtools": "^12.5.1",
-    "angular-in-memory-web-api": "^0.9.0",
+    "angular-in-memory-web-api": "^0.14.0",
     "moment": "^2.24.0",
-    "rxjs": "~6.5.5",
+    "rxjs": "~7.5.6",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.6"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1000.2",
     "@angular/cli": "~10.0.2",
-    "@angular/compiler-cli": "~12.2.16",
-    "@angular/language-service": "~12.2.16",
+    "@angular/compiler-cli": "~14.0.6",
+    "@angular/language-service": "~14.0.6",
     "@ngrx/schematics": "^12.5.1",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
@@ -52,6 +52,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.7.4"
   }
 }
