@@ -52,62 +52,61 @@ import {MatTooltipModule} from '@angular/material/tooltip';
 import {CanDeactivateGuard} from './directives/can-deactivate-guard/can-deactivate-guard';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    AdminLandingPageComponent,
-    AnnualExpenditureReportComponent,
-    AnnualExpenditureReportReconciliationComponent,
-    CapitalInvestmentPlanEditComponent,
-    CapitalInvestmentPlanViewComponent,
-    MunicipalCipContainerComponent,
-    MunicipalContactInformationComponent,
-    MunicipalLandingPageComponent,
-    MunicipalReportsComponent,
-    RegionalCipContainerComponent,
-    RegionalLandingPageComponent,
-  ],
-  imports: [
-    AppRoutingModule,
-    BrowserModule,
-    BrowserAnimationsModule,
-    ControlsModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatDatepickerModule,
-    MatDialogModule,
-    MatFormFieldModule,
-    MatGridListModule,
-    MatIconModule,
-    MatInputModule,
-    MatListModule,
-    MatMomentDateModule,
-    MatSelectModule,
-    MatSnackBarModule,
-    MatTableModule,
-    MatTabsModule,
-    MatTooltipModule,
-    ReactiveFormsModule,
-    StoreModule.forRoot({
-      'static-data': staticDataReducer,
-      'capital-investment-plan': capitalInvestmentPlanReducer,
-      'annual-expenditure-report': annualExpenditureReportReducer,
-      'contact-info': contactInfoReducer,
-      'municipal-allocations': municipalAllocationsReducer,
-    }),
-    EffectsModule.forRoot([
-      StaticDataEffects,
-      CIPEffects,
-      AEREffects,
-      ContactInfoEffects,
-      MunicipalAllocationsEffects,
-    ]),
-    !environment.production ? StoreDevtoolsModule.instrument() : [],
-    // StoreModule.forRoot({ cip: cipReducer }),
-    // StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production })
-  ],
-  providers: [CanDeactivateGuard],
-  bootstrap: [AppComponent],
-  entryComponents: [],
+    declarations: [
+        AppComponent,
+        AdminLandingPageComponent,
+        AnnualExpenditureReportComponent,
+        AnnualExpenditureReportReconciliationComponent,
+        CapitalInvestmentPlanEditComponent,
+        CapitalInvestmentPlanViewComponent,
+        MunicipalCipContainerComponent,
+        MunicipalContactInformationComponent,
+        MunicipalLandingPageComponent,
+        MunicipalReportsComponent,
+        RegionalCipContainerComponent,
+        RegionalLandingPageComponent,
+    ],
+    imports: [
+        AppRoutingModule,
+        BrowserModule,
+        BrowserAnimationsModule,
+        ControlsModule,
+        MatButtonModule,
+        MatCheckboxModule,
+        MatDatepickerModule,
+        MatDialogModule,
+        MatFormFieldModule,
+        MatGridListModule,
+        MatIconModule,
+        MatInputModule,
+        MatListModule,
+        MatMomentDateModule,
+        MatSelectModule,
+        MatSnackBarModule,
+        MatTableModule,
+        MatTabsModule,
+        MatTooltipModule,
+        ReactiveFormsModule,
+        StoreModule.forRoot({
+            'static-data': staticDataReducer,
+            'capital-investment-plan': capitalInvestmentPlanReducer,
+            'annual-expenditure-report': annualExpenditureReportReducer,
+            'contact-info': contactInfoReducer,
+            'municipal-allocations': municipalAllocationsReducer,
+        }),
+        EffectsModule.forRoot([
+            StaticDataEffects,
+            CIPEffects,
+            AEREffects,
+            ContactInfoEffects,
+            MunicipalAllocationsEffects,
+        ]),
+        !environment.production ? StoreDevtoolsModule.instrument() : [],
+        // StoreModule.forRoot({ cip: cipReducer }),
+        // StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production })
+    ],
+    providers: [CanDeactivateGuard],
+    bootstrap: [AppComponent]
 })
 export class AppModule {
 }
